+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Title bar appearance will no longer push webview down
+
+### Added
+- Added option to fit webview when title bar appears
 
 ## [0.2.0] - 2019-07-19
 ### Changed